Moved the StreamingXMLWriter inside node.py
