+++ openquake/commonlib/nrml_examples/source_model/alternative-mfds_4test.xml
@@ -62,7 +62,6 @@ xmlns:gml="http://www.opengis.net/gml"
                     60.0
                 </rake>
                 <surface>
-                    
                     <complexFaultGeometry>
                         
                         <faultTopEdge>
